Git: missing newline at end of "glo" alias
The pretty formatting was using the "format" format, which puts
separators *between* the log entries. Since the separator is a newline
by default, this means that all the log entries will have newlines at
the end of the lines, except for the last one.

This lack of a newline at the end of the last log entry will result in
problems such "wc -l" reporting one line too few, if you pipe the output
from "glo" into it.

Use the "tformat" format, since that puts the separator at the end of
each log entry, including the last one.