Port prompt_linduxed to prompt_sorin's structure
This patch does the following things:

1. Generally overwrites the structure of prompt_linduxed with the one of
   prompt_sorin.
2. Ports the general look of the current prompt_linduxed.
3. Makes some minor changes to the old look of prompt_linduxed.

The main reason for #1 is that the prompt_sorin code has being updated
and improved, while the code of prompt_linduxed hasn't been.
While the asynchronous updating of the git branch is the big addition,
other minor improvements are also worth importing into prompt_linduxed.

Keep most of the look of the current prompt, except for the following
changes:

* Make username and hostname only visible when ssh'd into a server.
* Add a timestamp at the start of the prompt.
* Remove the prefix "git:" from the branch name.