Replace blinking with bold text
The `environment` module seems to intend making certain text bold, but
it uses `;5` (which is the code for `BLINK`) instead of `;1` (which is
the code for `BOLD`).

I haven't noticed this until I recently moved to Konsole, where `BLINK`
is not disabled. Maybe the developers of Prezto use a terminal that
renders `BLINK` as `BOLD` (just like I did, until recently).